refactor(service): use env variables for configuration (#27)
This PR updates how the service reads the configuration. It was read from a "database" file that wasn't tracked with `git`. The deployment was tied to the history of the production branch which is not flexible.

Now we have a new environment variable called `HELP_SCOUT_MAILBOXES` which the configuration stringified. The process to add a configuration is straight forward: get the value, parse it, update it, encode it and finally restart the service.

To make this process easier to implement/read I've extracted the logic of the service into its own module. We have a new `main` module which does only the plumbing between the dependencies e.g. read/parse environment variable, create the HelpScout client, ... At the end it makes the service easier to test.